+++ lib/r_list/ruby.ex
@@ -11,37 +11,22 @@ defmodule RList.Ruby do
     RUtils.define_all_functions!(__MODULE__)
   end
 
-  use REnum.Ruby
-  import REnum.Support
-
   # https://ruby-doc.org/core-3.1.0/Array.html
   # [:all?, :any?, :append, :assoc, :at, :bsearch, :bsearch_index, :clear, :collect, :collect!, :combination, :compact, :compact!, :concat, :count, :cycle, :deconstruct, :delete, :delete_at, :delete_if, :difference, :dig, :drop, :drop_while, :each, :each_index, :empty?, :eql?, :fetch, :fill, :filter, :filter!, :find_index, :first, :flatten, :flatten!, :hash, :include?, :index, :initialize_copy, :insert, :inspect, :intersect?, :intersection, :join, :keep_if, :last, :length, :map, :map!, :max, :min, :minmax, :none?, :old_to_s, :one?, :pack, :permutation, :pop, :prepend, :product, :push, :rassoc, :reject, :reject!, :repeated_combination, :repeated_permutation, :replace, :reverse, :reverse!, :reverse_each, :rindex, :rotate, :rotate!, :sample, :select, :select!, :shift, :shuffle, :shuffle!, :size, :slice, :slice!, :sort, :sort!, :sort_by!, :sum, :take, :take_while, :to_a, :to_ary, :to_h, :to_s, :transpose, :union, :uniq, :uniq!, :unshift, :values_at, :zip]
-  # |> RUtils.required_functions([List, REnum.Ruby])
-  # ✔ all?
-  # ✔ any?
+  # |> RUtils.required_functions([List, REnum])
   # ✔ append
   # ✔ assoc
-  # ✔ at
   # bsearch
   # bsearch_index
   # ✔ clear
   # combination
-  # ✔ concat
-  # ✔ count
   # deconstruct
-  # ✔ delete_if
+  # delete_if
   # ✔ difference
-  # dig
-  # drop
-  # drop_while
-  # each
+  # ✔ dig
   # each_index
-  # empty?
   # eql?
-  # fetch
   # fill
-  # ✔ filter
-  # find_index
   # hash
   # index
   # initialize_copy
@@ -49,63 +34,29 @@ defmodule RList.Ruby do
   # inspect
   # intersect?
   # intersection
-  # join
-  # ✔ keep_if
+  # keep_if
   # length
-  # map
-  # max
-  # min
   # old_to_s
   # pack
   # permutation
   # pop
   # prepend
-  # product
   # push
   # rassoc
-  # ✔ reject
   # repeated_combination
   # repeated_permutation
   # replace
-  # reverse
   # rindex
   # rotate
   # sample
   # shift
-  # shuffle
   # size
-  # slice
-  # sort
-  # sum
-  # take
-  # take_while
   # to_ary
   # to_s
   # transpose
   # union
-  # uniq
   # unshift
   # values_at
-  def all?(list, function_or_pattern) do
-    cond do
-      is_function(function_or_pattern) ->
-        list |> Enum.all?(function_or_pattern)
-
-      true ->
-        list |> Enum.filter(match_function(function_or_pattern)) |> Enum.count() ==
-          Enum.count(list)
-    end
-  end
-
-  def any?(list, function_or_pattern) do
-    cond do
-      is_function(function_or_pattern) ->
-        list |> Enum.any?(function_or_pattern)
-
-      true ->
-        list |> Enum.filter(match_function(function_or_pattern)) |> Enum.count() > 0
-    end
-  end
 
   def push(list, elements_or_element) do
     list ++ List.wrap(elements_or_element)
@@ -120,11 +71,6 @@ defmodule RList.Ruby do
   end
 
   def clear(list) when is_list(list), do: []
-
-  def concat(list, lists) do
-    (list ++ List.wrap(lists))
-    |> List.flatten()
-  end
 
   def difference(list1, list2) do
     list1 -- list2
@@ -150,15 +96,5 @@ defmodule RList.Ruby do
   #   |> func.()
   # end
 
-  defdelegate all?(list), to: Enum
-  defdelegate any?(list), to: Enum
   defdelegate append(list, elements), to: __MODULE__, as: :push
-  defdelegate at(list, index), to: Enum
-  defdelegate map(list, func), to: Enum
-  defdelegate count(list), to: Enum
-  defdelegate count(list, func), to: Enum
-  defdelegate delete_if(list, func), to: Enum, as: :reject
-  defdelegate reject(list, func), to: Enum
-  defdelegate keep_if(list, func), to: Enum, as: :filter
-  defdelegate filter(list, func), to: Enum
 end