+++ lib/r_range.ex
@@ -1,10 +1,11 @@
 defmodule RRange do
   @moduledoc """
-  Entry point of Range extensions, and can use all of RRange.* functions.
+  Entry point of Range extensions, and can use all of RRange.* and REnum functions.
   See also.
    - [RRange.Native](https://hexdocs.pm/r_enum/RRange.Native.html#content)
   """
   use RRange.Native
   use RRange.Ruby
   use RRange.ActiveSupport
+  use REnum, undelegate_functions: Range.module_info()[:exports] |> Keyword.keys()
 end