+++ lib/r_stream.ex
@@ -1,10 +1,11 @@
 defmodule RStream do
   @moduledoc """
-  Entry point of Stream extensions, and can use all of RStream.* functions.
+  Entry point of Stream extensions, and can use all of RStream.* and REnum functions.
   See also.
    - [RStream.Native](https://hexdocs.pm/r_enum/RStream.Native.html#content)
   """
   use RStream.Native
   use RStream.Ruby
   use RStream.ActiveSupport
+  use REnum, undelegate_functions: Stream.module_info()[:exports] |> Keyword.keys()
 end