+++ CHANGELOG.md
@@ -4,9 +4,22 @@
 
 ## v0.4.0 (2022-1-xx)
 
+### Features
+
 - RList.Ruby
+- RList, Range, RMap and Stream has all REnum functions without duplicate native functions.
+
+### Breaking changes
+
+- Removes deprecated native functions.
 
 ## v0.3.2 (2022-1-20)
+
+### Features
+
+- Add `REnum.each_slice/2` that returns Stream.
+- Add alias `REnum.to_l` of `REnum.to_lilst`.
+- Add `mix dialyzer` to test.
 
 ### Breaking changes
 
@@ -16,10 +29,7 @@
 
 ### Other
 
-- Add `REnum.each_slice/2` that returns Stream.
-- Add `mix dialyzer` to test.
 - Fix hex version badge.
-- Add alias `REnum.to_l` of `REnum.to_lilst`.
 
 ## v0.3.1 (2022-1-17)